[simulator][machine] merge into one file
